Fixes togglePlay() unhandled exception in promise
Fixes the new autoplay restrictions where Play() can fail due to Chrome/Safari not registering a user interaction before the video plays (with sound). In this case, when a same-host iframe loads a call to video.play() is invoked to auto-start the video. If the user has been using the app then no problem the video autoplays fine. But if this user takes a link and sends it to one of their friends, who autoplays it when they load the link. In this scenario, chrome will actually block the video play and invoke an exception in a promise from video.play(). This promise is not correctly referenced and therefore the orphaned promise errors to the console instead of being handled.

The solution is to change togglePlay() such that the promise is returned to the caller, where the error condition can be gracefully handled.

This enables an attempt at autoplay, with a fallback behavior such as displaying a play.svg button.